Add Rules and move out of separate file in each example
