+++ examples/morpheusvm/genesis/genesis.go
@@ -8,8 +8,10 @@ import (
 	"encoding/json"
 	"fmt"
 
+	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/trace"
 
+	"github.com/ava-labs/hypersdk/chain"
 	"github.com/ava-labs/hypersdk/codec"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/storage"
@@ -65,4 +67,8 @@ func (g *Genesis) Load(ctx context.Context, tracer trace.Tracer, mu state.Mutabl
 		}
 	}
 	return nil
+}
+
+func (g *Genesis) Rules(t int64, networkID uint32, chainID ids.ID) chain.Rules {
+	return rules.New(&g.Genesis, networkID, chainID)
 }
