+++ examples/morpheusvm/tests/integration/integration_test.go
@@ -41,6 +41,7 @@ import (
 	"github.com/ava-labs/hypersdk/fees"
 	"github.com/ava-labs/hypersdk/pubsub"
 	"github.com/ava-labs/hypersdk/rpc"
+	"github.com/ava-labs/hypersdk/rules"
 	"github.com/ava-labs/hypersdk/vm"
 
 	auth "github.com/ava-labs/hypersdk/auth"
@@ -176,7 +177,7 @@ var _ = ginkgo.BeforeSuite(func() {
 	gen = genesis.Default()
 	gen.MinUnitPrice = fees.Dimensions{1, 1, 1, 1, 1}
 	gen.MinBlockGap = 0
-	gen.CustomAllocation = []*genesis.CustomAllocation{
+	gen.CustomAllocation = []*rules.CustomAllocation{
 		{
 			Address: addrStr,
 			Balance: 10_000_000,
